requirements.txt
pixoo
Pillow
psutil
numpy
soundcard
pynvml
opencv-python
feedparser
customtkinter
spotipy
icalendar
pycryptodome
lzallright